Automatically backed up by Learn

+++ lib/create.sql
@@ -2,6 +2,6 @@ CREATE TABLE projects(
   id INTEGER PRIMARY KEY,
   title TEXT,
   catefory TEXT,
-  funding_goal FLOAT(2),
+  funding_goal NUMERIC,
 
 );
